requirements.txt
python-docx==0.8.11
docx2python==2.6.3 